多签钱包demo

+++ test/nftMarket_study/TestNftMarketTest.sol
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.13;
 
-import {Test, console} from "forge-std/Test.sol";
-import {BaseERC20} from "../../src/nftMarket_study/baseERC20.sol";
-import {MyNFT} from "../../src/nftMarket_study/myNft.sol";
-import {NFTMarket} from "../../src/nftMarket_study/nftMarket.sol";
+import { Test, console } from "forge-std/Test.sol";
+import { BaseERC20 } from "../../src/nftMarket_study/baseERC20.sol";
+import { MyNFT } from "../../src/nftMarket_study/myNft.sol";
+import { NFTMarket } from "../../src/nftMarket_study/nftMarket.sol";
 
 //forge test --mc TestNftMarketTest
 contract TestNftMarketTest is Test {
@@ -27,7 +27,7 @@ contract TestNftMarketTest is Test {
     }
 
     mapping(uint256 => uint256) private tokenIds; //存nftId和价格
-    uint[] private tokenIdsArr; //存所有nftId
+    uint256[] private tokenIdsArr; //存所有nftId
 
     function test_list() public {
         address nftSeller = makeAddr("nftSeller"); // nft卖家
@@ -42,16 +42,11 @@ contract TestNftMarketTest is Test {
         myNFTInstance.mintNFT("https://www.myNFT.com/1");
         myNFTInstance.approve(address(nftMarketInstance), tokenId);
         //断言授权是否成功
-        assertEq(
-            myNFTInstance.getApproved(tokenId),
-            address(nftMarketInstance)
-        );
+        assertEq(myNFTInstance.getApproved(tokenId), address(nftMarketInstance));
         //上架nft设置价格
         nftMarketInstance.list(tokenId, nftPrice);
         //断言是否上架成功
-        NFTMarket.Listing memory listing = nftMarketInstance.getNftByTokenId(
-            tokenId
-        );
+        NFTMarket.Listing memory listing = nftMarketInstance.getNftByTokenId(tokenId);
         assertEq(listing.seller, nftSeller);
         assertEq(listing.price, nftPrice);
         vm.stopPrank();
@@ -69,36 +64,21 @@ contract TestNftMarketTest is Test {
         assertEq(baseERC20Instance.balanceOf(nftBuyer), nftToBuyUserAmount);
         //断言低于上市价格购买 判断错误是否等于
         vm.expectRevert("Insufficient amount"); //断言错误的方法
-        baseERC20Instance.transferWithCallback(
-            address(nftMarketInstance),
-            nftPrice - 1,
-            tokenId
-        );
+        baseERC20Instance.transferWithCallback(address(nftMarketInstance), nftPrice - 1, tokenId);
 
         //买家付款给交易所
-        baseERC20Instance.transferWithCallback(
-            address(nftMarketInstance),
-            nftPrice,
-            tokenId
-        );
+        baseERC20Instance.transferWithCallback(address(nftMarketInstance), nftPrice, tokenId);
         //检查卖家是否收到款项
         assertEq(baseERC20Instance.balanceOf(nftSeller), nftPrice);
         //检查nft是否已经到了买家
         assertEq(myNFTInstance.ownerOf(tokenId), nftBuyer);
 
         //检查买家余额是否正确
-        assertEq(
-            baseERC20Instance.balanceOf(nftBuyer),
-            nftToBuyUserAmount - nftPrice
-        );
+        assertEq(baseERC20Instance.balanceOf(nftBuyer), nftToBuyUserAmount - nftPrice);
 
         //断言购买一个不存在的nft
         vm.expectRevert("does not exist");
-        baseERC20Instance.transferWithCallback(
-            address(nftMarketInstance),
-            nftPrice,
-            tokenId
-        );
+        baseERC20Instance.transferWithCallback(address(nftMarketInstance), nftPrice, tokenId);
         //随机测试
         vm.startPrank(nftSeller);
         for (uint256 i = 0; i < 10; i++) {
@@ -121,19 +101,14 @@ contract TestNftMarketTest is Test {
             vm.startPrank(nftBuyer2);
             tokenId = tokenIdsArr[i];
             uint256 nprice = tokenIds[tokenId];
-            baseERC20Instance.transferWithCallback(
-                address(nftMarketInstance),
-                nprice,
-                tokenId
-            );
+            baseERC20Instance.transferWithCallback(address(nftMarketInstance), nprice, tokenId);
             assertEq(myNFTInstance.ownerOf(tokenId), nftBuyer2);
             vm.stopPrank();
         }
     }
 
     function randomUrl(uint256 seed) private view returns (string memory) {
-        bytes
-            memory alphabet = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+        bytes memory alphabet = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
         bytes memory url = new bytes(16); // 生成长度为16的随机字符串
         for (uint256 i = 0; i < 16; i++) {
             url[i] = alphabet[random(seed + i) % alphabet.length];
@@ -142,8 +117,7 @@ contract TestNftMarketTest is Test {
     }
 
     function randomName(uint256 seed) private view returns (string memory) {
-        bytes
-            memory alphabet = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+        bytes memory alphabet = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
         bytes memory name = new bytes(4); // 生成长度为16的随机字符串
         for (uint256 i = 0; i < 4; i++) {
             name[i] = alphabet[random(seed + i) % alphabet.length];
@@ -152,8 +126,6 @@ contract TestNftMarketTest is Test {
     }
 
     function random(uint256 seed) private view returns (uint256) {
-        return
-            (uint256(keccak256(abi.encodePacked(block.timestamp, seed))) %
-                100) + 1;
+        return (uint256(keccak256(abi.encodePacked(block.timestamp, seed))) % 100) + 1;
     }
 }
